requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
cmake==3.25.2
colorzero==2.0
Django==4.1.6
djangorestframework==3.14.0
gpiozero==1.6.2
idna==3.4
jet==1.0
MQ2==1.1.0
Pillow==9.4.0
pytz==2022.7.1
requests==2.28
sqlparse==0.4.3
straight.plugin==1.5.0
urllib3==1.26.14
xlrd==2.0.1
